+++ src/Controller/PlanningController.php
@@ -12,11 +12,9 @@ use App\Service\DataProviderService;
 use App\Service\ViewService;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/admin/planning')]
 class PlanningController extends AbstractController
@@ -24,20 +22,52 @@ class PlanningController extends AbstractController
     public function __construct(
         private readonly DataProviderService $dataProviderService,
         private readonly DataProviderRepository $dataProviderRepository,
-        private readonly TranslatorInterface $translator,
         private readonly ViewService $viewService,
+        private readonly ?string $planningDefaultDataProvider,
     ) {
+    }
+
+    #[Route('/', name: 'app_planning')]
+    public function index(Request $request): Response
+    {
+        return $this->redirectToRoute('app_planning_users');
     }
 
     /**
      * @throws UnsupportedDataProviderException
      * @throws EconomicsException
      */
-    #[Route('/', name: 'app_planning')]
-    public function index(Request $request): Response
+    #[Route('/users', name: 'app_planning_users')]
+    public function planningUsers(Request $request): Response
+    {
+        return $this->createResponse($request, 'users');
+    }
+
+    /**
+     * @throws UnsupportedDataProviderException
+     * @throws EconomicsException
+     */
+    #[Route('/projects', name: 'app_planning_projects')]
+    public function planningProjects(Request $request): Response
+    {
+        return $this->createResponse($request, 'projects');
+    }
+
+    /**
+     * @throws UnsupportedDataProviderException
+     * @throws EconomicsException
+     */
+    private function createResponse(Request $request, string $mode): Response
     {
         $planningFormData = new PlanningFormData();
         $form = $this->createForm(PlanningType::class, $planningFormData);
+
+        $dataProviders = $this->dataProviderRepository->findAll();
+        $defaultProvider = $this->dataProviderRepository->find($this->planningDefaultDataProvider);
+
+        if (null === $defaultProvider && count($dataProviders) > 0) {
+            $defaultProvider = $dataProviders[0];
+        }
 
         $form->add('dataProvider', EntityType::class, [
             'class' => DataProvider::class,
@@ -48,56 +78,33 @@ class PlanningController extends AbstractController
                 'class' => 'form-element',
             ],
             'help' => 'planning.data_provider_helptext',
-            'choices' => $this->dataProviderRepository->findAll(),
-        ]);
-
-        $form->add('viewType', ChoiceType::class, [
-            'required' => true,
-            'label' => 'planning.view_type',
-            'label_attr' => ['class' => 'label'],
-            'attr' => [
-                'class' => 'form-element',
-            ],
-            'help' => 'planning.data_provider_helptext',
-            'choices' => $this->getTypeChoices(),
+            'data' => $this->dataProviderRepository->find($this->planningDefaultDataProvider),
+            'choices' => $dataProviders,
         ]);
 
         $form->handleRequest($request);
 
-        $planningData = null;
-        $template = 'planning/index.html.twig';
+        $service = null;
 
         if ($form->isSubmitted() && $form->isValid()) {
             $service = $this->dataProviderService->getService($planningFormData->dataProvider);
-            $viewType = $form->getData()->viewType;
-
-            switch ($viewType) {
-                case 'week':
-                    $planningData = $service->getPlanningDataWeeks();
-                    $template = 'planning/planning-weeks.html.twig';
-                    break;
-                case 'sprint':
-                    $planningData = $service->getPlanningDataSprints();
-                    $template = 'planning/planning-sprints.html.twig';
-                    break;
-                default:
-                    $planningData = $service->getPlanningDataSprints();
-                    break;
-            }
+        } elseif (null !== $defaultProvider) {
+            $service = $this->dataProviderService->getService($defaultProvider);
         }
 
-        return $this->render($template, $this->viewService->addView([
+        try {
+            $planningData = $service?->getPlanningDataWeeks();
+        } catch (\Exception $e) {
+            $error = $e->getMessage();
+            $planningData = null;
+        }
+
+        return $this->render('planning/planning.html.twig', $this->viewService->addView([
             'controller_name' => 'PlanningController',
             'planningData' => $planningData,
+            'error' => $error ?? null,
             'form' => $form,
+            'mode' => $mode,
         ]));
-    }
-
-    private function getTypeChoices()
-    {
-        return [
-            $this->translator->trans('planning.week_view') => 'week',
-            $this->translator->trans('planning.sprint_view') => 'sprint',
-        ];
     }
 }