+++ src/Form/InvoiceRecordType.php
@@ -13,13 +13,14 @@ class InvoiceRecordType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('confirmed', ChoiceType::class, [
+            ->add('confirmation', ChoiceType::class, [
                 'required' => true,
                 'label' => 'invoices.record_invoice',
                 'label_attr' => ['class' => 'label'],
                 'choices' => [
-                    'invoices.record_invoice_false' => false,
-                    'invoices.record_invoice_true' => true,
+                    'invoices.record_invoice_no' => ConfirmData::INVOICE_RECORD_NO,
+                    'invoices.record_invoice_yes' => ConfirmData::INVOICE_RECORD_YES,
+                    'invoices.record_invoice_yes_no_cost' => ConfirmData::INVOICE_RECORD_YES_NO_COST,
                 ],
                 'help' => 'invoices.record_invoice_helptext',
                 'attr' => ['class' => 'form-element'],
@@ -30,7 +31,6 @@ class InvoiceRecordType extends AbstractType
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'method' => 'GET',
             'data_class' => ConfirmData::class,
         ]);
     }