+++ src/Model/Invoices/InvoiceFilterData.php
@@ -4,7 +4,9 @@ namespace App\Model\Invoices;
 
 class InvoiceFilterData
 {
+    public ?string $query = '';
     public ?bool $recorded = false;
+    public ?bool $noCost = null;
     public ?string $createdBy = '';
     public ?bool $projectBilling = false;
 }